Updated board file; started cover CAD.
This checkin marks the hopefully complete V1.0 design for the form-factor
board.

It also contains the first (incomplete) iteration of the cover design.
It has a bunch of interference with components on the PCB which
needs to be cleaned. It's also unclear if snaps have to be 1.5mm thick.
Seems excessive. Finally, I don't think the OLED display can even survive
such stiff snaps. But it at least a start.