Added new OLED display "datasheet".
This should be more or less functionally compatible to the
old one, but much more widely available and cheaper (on AliBaba), being the form factor that's used in the Ledger
nano S thingy.

It is configured as 4-wire SPI (i.e. separate C/D pin), which is a
bit of a bummer, but what can you do.